\name{ShipWeight}
\alias{ShipWeight}
\title{Ship weight data for linear regression}
\description{
    Weights of ships with males, females, cats, and ants on them.
}
\usage{data(ShipWeight)}
\format{
  \tabular{ll}{
    Weight_Kg:\tab Weight of each ship\cr
    Male:\tab Number of males on each ship\cr
    Female:\tab Number of females on each ship\cr
    Cat:\tab Number of cats on each ship\cr
    Ant:\tab Number of ants on each ship\cr
  }
}
\note{
  Sample file for linear regression.
}

\keyword{datasets}
\keyword{RcmdrPlugin.EZR}

